# tests_data/openapi_pre_destructure.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelNet API",
    "version": "0.2.0-beta.1"
  },
  "paths": {
...
    "/api/provisioner/generate": {
      "post": {
        "tags": [
          "Provisioner"
        ],
        "summary": "Provisioner Generate",
        "description": "Genera la running-config e la restituisce come testo (view/copy nella UI).\nDi default i segreti sono placeholder; ``?materialized=true`` per il testo\ncompleto (auditato).",
        "operationId": "provisioner_generate_api_provisioner_generate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "materialized",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Materialized"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchProvisionSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/provisioner/download": {
      "post": {
        "tags": [
          "Provisioner"
        ],
        "summary": "Provisioner Download",
        "description": "Genera la running-config e la restituisce come file .txt scaricabile.",
        "operationId": "provisioner_download_api_provisioner_download_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "materialized",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Materialized"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchProvisionSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/provisioner/push-ssh": {
      "post": {
        "tags": [
          "Provisioner"
        ],
        "summary": "Provisioner Push Ssh",
        "description": "Genera la config e la applica via SSH (Netmiko) su un apparato raggiungibile.",
        "operationId": "provisioner_push_ssh_api_provisioner_push_ssh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchProvisionSSHSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/provisioner/push-serial": {
      "post": {
        "tags": [
          "Provisioner"
        ],
        "summary": "Provisioner Push Serial",
        "description": "Genera la config e la applica via console/seriale (pyserial) per il\nprovisioning day-0 senza connettivita' di rete.",
        "operationId": "provisioner_push_serial_api_provisioner_push_serial_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchProvisionSerialSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/provisioner/serial-ports": {
      "get": {
        "tags": [
          "Provisioner"
        ],
        "summary": "Provisioner Serial Ports",
        "description": "Elenca le porte COM/seriali disponibili sull'host del server.",
        "operationId": "provisioner_serial_ports_api_provisioner_serial_ports_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/provisioner/fgt/generate": {
      "post": {
        "tags": [
          "Provisioner"
        ],
        "summary": "Fgt Provisioner Generate",
        "description": "Genera la configurazione FortiOS day-0 e la restituisce come testo.",
        "operationId": "fgt_provisioner_generate_api_provisioner_fgt_generate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "materialized",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Materialized"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FortiGateProvisionSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/provisioner/fgt/download": {
      "post": {
        "tags": [
          "Provisioner"
        ],
        "summary": "Fgt Provisioner Download",
        "description": "Genera la configurazione FortiOS e la restituisce come file .txt.",
        "operationId": "fgt_provisioner_download_api_provisioner_fgt_download_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "materialized",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Materialized"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FortiGateProvisionSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/provisioner/fgt/push-ssh": {
      "post": {
        "tags": [
          "Provisioner"
        ],
        "summary": "Fgt Provisioner Push Ssh",
        "description": "Genera la config FortiOS e la applica sul device: REST API prima (se un\ntoken e' salvato per l'host, stesso pattern dell'osservabilità), con\nfallback SSH (Netmiko 'fortinet'). 'method' nel risultato indica il canale\nusato; 'api_error' spiega l'eventuale fallback.",
        "operationId": "fgt_provisioner_push_ssh_api_provisioner_fgt_push_ssh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FortiGateProvisionSSHSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/provisioner/fgt/push-serial": {
      "post": {
        "tags": [
          "Provisioner"
        ],
        "summary": "Fgt Provisioner Push Serial",
        "description": "Genera la config FortiOS e la applica via console/seriale (day-0).",
        "operationId": "fgt_provisioner_push_serial_api_provisioner_fgt_push_serial_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FortiGateProvisionSerialSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/identities": {
      "get": {
        "tags": [
          "Provisioner"
        ],
        "summary": "Identities List",
        "description": "Lista identita' (senza segreti), opzionalmente filtrate per tenant.",
        "operationId": "identities_list_api_identities_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Provisioner"
        ],
        "summary": "Identities Create",
        "operationId": "identities_create_api_identities_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentitySchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/identities/{identity_id}": {
      "put": {
        "tags": [
          "Provisioner"
        ],
        "summary": "Identities Update",
        "operationId": "identities_update_api_identities__identity_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentitySchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Provisioner"
        ],
        "summary": "Identities Delete",
        "operationId": "identities_delete_api_identities__identity_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IdentitySchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tenant": {
            "type": "string",
            "title": "Tenant"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "enable_secret": {
            "type": "string",
            "title": "Enable Secret",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name",
          "tenant",
          "username",
          "password"
        ],
        "title": "IdentitySchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}